pyproject.toml
[project]
name = "spotify-snapshot"
description = "A tool to backup Spotify library data to Git"
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dynamic = ["version"]
authors = [
	{ name = "Patrick Riggs", email = "riggspc@gmail.com" },
	{ name = "Aaron Lichtman", email = "aaronlichtman@gmail.com" }
]
keywords = ["spotify", "backup", "git"]
dependencies = [
	"loguru",
	"click",
	"rich",
	"python-crontab",
	"spotipy",
	"gitpython",
	"tomli",
	"tomli-w",
	"requests",
]

[project.urls]
"Source Code" = "https://github.com/alichtman/spotify-snapshot"
"Issue Tracker" = "https://github.com/alichtman/spotify-snapshot/issues"


[project.scripts]
spotify-snapshot = "spotify_snapshot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "spotify_snapshot/__about__.py"

[tool.hatch.build]
include = ["spotify_snapshot/*"]
artifacts = ["dist/*"]

[tool.mypy]
# Prevent duplicate module checking
exclude = ["build/", "dist/"]
# Ignore missing imports for external packages
ignore_missing_imports = true
# Make the checker stricter
strict = true
# Don't check the same file twice
explicit_package_bases = true
namespace_packages = true
# The root directory to check
mypy_path = "."

[tool.hatch.envs.default]
dependencies = [
	# Development dependencies that were previously mixed with runtime dependencies
	"rope",
	"better-exceptions",
	"black",
	"flake8",
	"mypy",
]

[tool.hatch.envs.default.scripts]
fmt = "black *.py spotify_snapshot/*.py"
lint = [
    "black --check *.py spotify_snapshot/*.py",
    "flake8 .",
    "mypy spotify_snapshot"
]
typecheck = "mypy spotify_snapshot"

[tool.hatch.envs.test]
dependencies = [
	"pytest",
	"pytest-cov",
]

[tool.hatch.envs.lint]
dependencies = [
	"black",
	"flake8",
	"mypy",
]

[tool.hatch.envs.lint.scripts]
style = [
	"black .",
	"flake8 .",
	"mypy .",
]
